S3 media storage and other improvements and fixes

+++ HCms.Application/Services/MediaManagement.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.StaticFiles;
+using Microsoft.Extensions.DependencyInjection;
 
 using AleProjects.Base64;
 using HCms.Application.Dto;
@@ -18,35 +19,103 @@ using HCms.Infrastructure.Notification;
 namespace HCms.Application.Services
 {
 
-	public class MediaManagementService(IMediaStorage mediaStorage, IAuthorizationService authService, IEventNotifier notifier)
+	public class MediaManagementService(
+		[FromKeyedServices("local")] IMediaStorage localMediaStorage,
+		[FromKeyedServices("s3")] IMediaStorage s3MediaStorage,
+		IAuthorizationService authService, 
+		IEventNotifier notifier)
 	{
-		private readonly IMediaStorage _mediaStorage = mediaStorage;
+		private readonly IMediaStorage _localMediaStorage = localMediaStorage;
+		private readonly IMediaStorage _s3MediaStorage = s3MediaStorage;
 		private readonly IAuthorizationService _authService = authService;
 		private readonly IEventNotifier _notifier = notifier;
 
+		static bool IsValidPath(string path) => string.IsNullOrEmpty(path) || !path.Contains("..");
 
-		public Result<DtoMediaFolderReadResult> Read(string link)
+		bool TryGetStorage(string path, out IMediaStorage storage)
 		{
-			if (!Base64Url.TryDecode(link, out string path))
+			if (_localMediaStorage.ServesPath(path))
+			{
+				storage = _localMediaStorage;
+				return true;
+			}
+
+			if (_s3MediaStorage.ServesPath(path))
+			{
+				storage = _s3MediaStorage;
+				return true;
+			}
+
+			storage = null;
+			return false;
+		}
+
+		public CommonMediaStorageParams GetCommonParams(string link)
+		{
+			if (!Base64Url.TryDecode(link.AsSpan(), out string path))
+				return CommonMediaStorageParams.Default();
+
+			if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
+				return CommonMediaStorageParams.Default();
+
+			return _mediaStorage.GetCommonParams(path);
+		}
+
+		public string GetDefaultPlace()
+		{
+			string[] keys = [.. _localMediaStorage.PlaceKeys, .. _s3MediaStorage.PlaceKeys];
+
+			return keys.FirstOrDefault();
+		}
+
+		public string GetDefaultDisplayPlace()
+		{
+			string[] keys = [.. _localMediaStorage.PlaceKeys, .. _s3MediaStorage.PlaceKeys];
+
+			return keys.Length == 1 ? keys[0] : null;
+		}
+
+		public async Task<Result<DtoMediaFolderReadResult>> Read(string link)
+		{
+			if (!Base64Url.TryDecode(link.AsSpan(), out string path))
 				return Result<DtoMediaFolderReadResult>.BadParameters("Link", "Invalid base64-url format");
 
-			if (!_mediaStorage.IsValidPath(path))
+			if (!IsValidPath(path))
 				return Result<DtoMediaFolderReadResult>.BadParameters("Link", "Invalid path");
 
-			var entries = _mediaStorage.ReadDirectory(path);
-
-			if (entries == null)
-				return Result<DtoMediaFolderReadResult>.NotFound();
-
 			DtoMediaStoragePathElement[] breadcrumbs;
+			List<Domain.Types.MediaStorageEntry> entries;
 
 			if (string.IsNullOrEmpty(path))
 			{
-				breadcrumbs = [new DtoMediaStoragePathElement()];
+				string[] keys = [.._localMediaStorage.PlaceKeys, .._s3MediaStorage.PlaceKeys];
+				entries = new(keys.Length);
+
+				foreach (string key in keys.Order())
+					entries.Add(new Domain.Types.MediaStorageEntry()
+					{
+						IsFolder = true,
+						Name = key,
+						FullName = key,
+						RelativeName = key,
+						Extension = string.Empty,
+						Date = DateTime.UnixEpoch
+					});
+
+				breadcrumbs = [new DtoMediaStoragePathElement() { Link = string.Empty }];
 			}
 			else
 			{
-				string[] labels = path.Split(Path.DirectorySeparatorChar);
+				if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
+					return Result<DtoMediaFolderReadResult>.BadParameters("Link", "Invalid path");
+
+				entries = await _mediaStorage.ReadDirectory(path);
+
+				if (entries == null)
+					return Result<DtoMediaFolderReadResult>.NotFound();
+
+
+				string[] labels = path.Split('/');
 				int n = labels.Length;
 				breadcrumbs = new DtoMediaStoragePathElement[n + 1];
 
@@ -57,9 +126,8 @@ namespace HCms.Application.Services
 					breadcrumbs[i + 1] = new DtoMediaStoragePathElement()
 					{
 						Label = labels[i],
-						Link = i == n - 1 ? null : Base64Url.Encode(string.Join(Path.DirectorySeparatorChar, labels, 0, i + 1))
+						Link = i == n - 1 ? null : Base64Url.Encode(string.Join('/', labels, 0, i + 1))
 					};
-
 				}
 			}
 
@@ -71,36 +139,23 @@ namespace HCms.Application.Services
 				});
 		}
 
-		public Result<DtoPhysicalMediaFileResult> Get(string link)
+		public async Task<Result<DtoPhysicalMediaFileResult>> Get(string link)
 		{
-			if (!Base64Url.TryDecode(link, out string path))
+			if (!Base64Url.TryDecode(link.AsSpan(), out string path))
 				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid base64-url format");
 
-			if (!_mediaStorage.IsValidPath(path))
+			if (!IsValidPath(path))
 				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid path");
 
-			var entry = _mediaStorage.GetFile(path);
+			if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
+				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid path");
+
+			var entry = await _mediaStorage.GetFile(path);
 
 			if (entry == null)
 				return Result<DtoPhysicalMediaFileResult>.NotFound();
 
 			return Result<DtoPhysicalMediaFileResult>.Success(new(entry));
-		}
-
-		public async Task<Result<DtoMediaStorageEntry>> Properties(string link)
-		{
-			if (!Base64Url.TryDecode(link, out string path))
-				return Result<DtoMediaStorageEntry>.BadParameters("Link", "Invalid base64-url format");
-
-			if (!_mediaStorage.IsValidPath(path))
-				return Result<DtoMediaStorageEntry>.BadParameters("Link", "Invalid path");
-
-			var entry = await _mediaStorage.Properties(path);
-
-			if (entry == null)
-				return Result<DtoMediaStorageEntry>.NotFound();
-
-			return Result<DtoMediaStorageEntry>.Success(new(entry));
 		}
 
 		public async Task<Result<DtoPhysicalMediaFileResult>> Preview(string link, int? size)
@@ -110,10 +165,13 @@ namespace HCms.Application.Services
 			else if (size <= 0)
 				return Result<DtoPhysicalMediaFileResult>.BadParameters("Size", "Must be positive");
 
-			if (!Base64Url.TryDecode(link, out string path))
+			if (!Base64Url.TryDecode(link.AsSpan(), out string path))
 				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid base64-url format");
 
-			if (!_mediaStorage.IsValidPath(path))
+			if (!IsValidPath(path))
+				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid path");
+
+			if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
 				return Result<DtoPhysicalMediaFileResult>.BadParameters("Link", "Invalid path");
 
 			var entry = await _mediaStorage.Preview(path, link, size.Value);
@@ -124,12 +182,37 @@ namespace HCms.Application.Services
 			return Result<DtoPhysicalMediaFileResult>.Success(new(entry));
 		}
 
+		public async Task<Result<DtoMediaStorageEntry>> Properties(string link)
+		{
+			if (!Base64Url.TryDecode(link.AsSpan(), out string path))
+				return Result<DtoMediaStorageEntry>.BadParameters("Link", "Invalid base64-url format");
+
+			if (!IsValidPath(path))
+				return Result<DtoMediaStorageEntry>.BadParameters("Link", "Invalid path");
+
+			if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
+				return Result<DtoMediaStorageEntry>.BadParameters("Link", "Invalid path");
+
+			var entry = await _mediaStorage.Properties(path);
+
+			if (entry == null)
+				return Result<DtoMediaStorageEntry>.NotFound();
+
+			return Result<DtoMediaStorageEntry>.Success(new(entry));
+		}
+
 		public async Task<Result<DtoMediaStorageEntry>> Save(Stream stream, string fileName, string destinationLink, ClaimsPrincipal user)
 		{
-			if (!Base64Url.TryDecode(destinationLink, out string destination))
+			if (!Base64Url.TryDecode(destinationLink.AsSpan(), out string destination))
 				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid base64-url format");
 
-			if (!_mediaStorage.IsValidPath(destination))
+			if (string.IsNullOrEmpty(destination))
+				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Media library root is read-only");
+
+			if (!IsValidPath(destination))
+				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid destination");
+
+			if (!TryGetStorage(destination, out IMediaStorage _mediaStorage))
 				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid destination");
 
 			var authResult = await _authService.AuthorizeAsync(user, "UploadUnsafeContent");
@@ -142,16 +225,18 @@ namespace HCms.Application.Services
 					!contentType.StartsWith("image/") && !contentType.StartsWith("video/"))
 					return Result<DtoMediaStorageEntry>.BadParameters(fileName, "Invalid file type");
 
-				if (!Regex.IsMatch(fileName, _mediaStorage.Settings.SafeNameRegex))
+				string safeNameRegex = _mediaStorage.GetCommonParams(destination).SafeNameRegex;
+
+				if (!string.IsNullOrEmpty(safeNameRegex) && !Regex.IsMatch(fileName, safeNameRegex))
 					return Result<DtoMediaStorageEntry>.BadParameters(fileName, "Unsafe file name");
 			}
 
 			var entry = await _mediaStorage.Save(stream, fileName, destination);
 
 			if (entry == null)
-				return Result<DtoMediaStorageEntry>.BadParameters(fileName, "Failed to save file");
+				return Result<DtoMediaStorageEntry>.BadParameters(fileName, "Failed to save a file");
 
-			await _notifier.Notify("on_media_create", [entry.FullName]);
+			await _notifier.Notify("on_media_create", [entry.RelativeName]);
 
 			return Result<DtoMediaStorageEntry>.Success(new(entry));
 		}
@@ -164,13 +249,22 @@ namespace HCms.Application.Services
 			string[] entries = new string[links.Length];
 
 			for (int i = 0; i < links.Length; i++)
-				if (Base64Url.TryDecode(links[i], out string path))
-					if (_mediaStorage.IsValidPath(path))
+				if (Base64Url.TryDecode(links[i].AsSpan(), out string path))
+					if (IsValidPath(path))
 						entries[i] = path;
 					else
 						return Result<string[]>.BadParameters(links[i], "Invalid path");
 				else
-						return Result<string[]>.BadParameters(links[i], "Invalid base64 format");
+					return Result<string[]>.BadParameters(links[i], "Invalid base64 format");
+
+			if (entries.Any(e => string.IsNullOrEmpty(Path.GetDirectoryName(e))))
+				return Result<string[]>.BadParameters("Links", "Media library root is read-only");
+
+			if (entries.Select(e => Path.GetDirectoryName(e)).Distinct().Count() > 1)
+				return Result<string[]>.BadParameters("Links", "All entries must be in the same location");
+
+			if (!TryGetStorage(entries.First(), out IMediaStorage _mediaStorage))
+				return Result<string[]>.BadParameters("Links", "Invalid path");
 
 			var list = await _mediaStorage.Delete(entries);
 			var deleted = list.Select(Base64Url.Encode).ToArray();
@@ -182,13 +276,18 @@ namespace HCms.Application.Services
 
 		public async Task<Result<DtoMediaStorageEntry>> CreateFolder(string name, string destination, ClaimsPrincipal user)
 		{
-			string path;
-
-			if (!Base64Url.TryDecode(destination, out path))
+			if (!Base64Url.TryDecode(destination.AsSpan(), out string path))
 				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid base64-url format");
 
-			if (!_mediaStorage.IsValidPath(path))
+			if (string.IsNullOrEmpty(path))
+				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Media library root is read-only");
+
+			if (!IsValidPath(path))
 				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid destination");
+
+			if (!TryGetStorage(path, out IMediaStorage _mediaStorage))
+				return Result<DtoMediaStorageEntry>.BadParameters("Destination", "Invalid destination");
+
 
 			var authResult = await _authService.AuthorizeAsync(user, "UploadUnsafeContent");
 
@@ -197,8 +296,15 @@ namespace HCms.Application.Services
 				if (!Regex.IsMatch(name, "^[\\w-]+$"))
 					return Result<DtoMediaStorageEntry>.BadParameters("Name", "Unsafe folder name");
 			}
+			else if (name.Any(c => c == '/' || c == '\\'))
+			{
+				return Result<DtoMediaStorageEntry>.BadParameters("Name", "Unsafe folder name");
+			}
 
-			var entry = _mediaStorage.CreateFolder(name, path);
+			var entry = await _mediaStorage.CreateFolder(name, path);
+
+			if (entry == null)
+				return Result<DtoMediaStorageEntry>.Other("Failed to create folder");
 
 			return Result<DtoMediaStorageEntry>.Success(new(entry));
 		}
